+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 ## [Unreleased]
 
 ### Added
+- **Breaking** Added `isLocalTile` to constructor of `DefaultVideoTile`, `VideoTileFactory` and `VideoTileState`
 * Added icons for the demo app
 * Added `SplashActivity` in demo app for displaying the launch screen
 * Added choose device list in the demo app for speaker button
 * Added metrics table in the demo app
 * Added `onVideoTileSizeChanged` API in `VideoTileObserver` for video stream content size change
+* Added attendee id to local video tile
 
 ### Changed
 * Changed the theme to support dark theme